feat: change uint8_t ip to uint16_t with other vars. Add new testcase

+++ chunk.c
@@ -11,17 +11,17 @@ void init_chunk(Chunk *chunk) {
 }
 
 void free_chunk(Chunk *chunk) {
-  FREE_ARRAY(uint8_t, chunk->code, chunk->capacity);
+  FREE_ARRAY(uint16_t, chunk->code, chunk->capacity);
   FREE_ARRAY(int, chunk->lines, chunk->capacity);
   free_value_array(&chunk->constants);
   init_chunk(chunk);
 }
 
-void write_chunk(Chunk *chunk, const uint8_t byte, const int line) {
+void write_chunk(Chunk *chunk, const uint16_t byte, const int line) {
   if (chunk->capacity < chunk->length + 1) {
     const int old_capacity = chunk->capacity;
     chunk->capacity = GROW_CAPACITY(old_capacity);
-    chunk->code = GROW_ARRAY(uint8_t, chunk->code, old_capacity, chunk->capacity);
+    chunk->code = GROW_ARRAY(uint16_t, chunk->code, old_capacity, chunk->capacity);
     chunk->lines = GROW_ARRAY(int, chunk->lines, old_capacity, chunk->capacity);
   }
